Reverse 16 CFileLoader functions (#96)
* Add CTimeModelInfo::FindOtherTimeModel

* Rename field tTimeInfo.m_nOtherTimeModel

* Fix signature CCullZones::AddMirrorAttributeZone

cm int -> float
center const CVector*  -> const CVector&

* Fix signature CCullZones::AddCullZone

center const CVector* -> const CVector&

* Add CGarages::AddOne

* Reverse 16 CFileLoader functions

* Add description CCullZoneReflection

* Use enum instead of magic number

rwID_CLUMP is better than 16


* Fix CAETwinLoopSoundEntity::PlayTwinLoopSound

(cherry picked from commit b1f9f3bf7234684386d2b5de8322782614d3ab33)

* Improve readability CFileLoader::LoadClumpFile


* Use std types in eGarageType and eGarageDoorState

* type fixes

+++ source/game_sa/CTimer.h
@@ -62,6 +62,6 @@ public:
     static void UpdateVariables(float timeStep);
     static void UpdateTimeStep(float fTimeStep);
 
-    static inline unsigned int GetTimeStepInMilliseconds() { return ms_fTimeStep * 20.0F; }
+    static inline uint32_t GetTimeStepInMilliseconds() { return static_cast<uint32_t>(ms_fTimeStep * 20.0F); }
     static void Update();
 };
